refactor: 컴포넌트별 폴더 구조화 및 index.tsx, constants.ts 분리 (#29)
- 아토믹 디자인 컨셉 참고하여 역할별/재사용성 기준으로 컴포넌트 구조화
- 각 컴포넌트별 index.tsx, constants.ts 등 역할별로 파일 정리
